Moved providers to their own namespace
