+++ Forge.Tests/Helpers/TestEventPayloadProvider.cs
@@ -1,6 +1,7 @@
 // Copyright © Gamesmiths Guild.
 
 using Gamesmiths.Forge.Statescript;
+using Gamesmiths.Forge.Statescript.Providers;
 
 namespace Gamesmiths.Forge.Tests.Helpers;
