+++ Forge.Tests/Statescript/Resolvers/CueCustomParametersResolverTests.cs
@@ -4,6 +4,7 @@ using FluentAssertions;
 using Gamesmiths.Forge.Core;
 using Gamesmiths.Forge.Statescript;
 using Gamesmiths.Forge.Statescript.Properties;
+using Gamesmiths.Forge.Statescript.Providers;
 
 namespace Gamesmiths.Forge.Tests.Statescript.Resolvers;
