+++ Forge.Tests/Statescript/Resolvers/EffectContextDataResolverTests.cs
@@ -5,6 +5,7 @@ using FluentAssertions;
 using Gamesmiths.Forge.Effects;
 using Gamesmiths.Forge.Statescript;
 using Gamesmiths.Forge.Statescript.Properties;
+using Gamesmiths.Forge.Statescript.Providers;
 
 namespace Gamesmiths.Forge.Tests.Statescript.Resolvers;
