+++ Forge/Statescript/Properties/EffectContextDataResolver.cs
@@ -1,6 +1,7 @@
 // Copyright © Gamesmiths Guild.
 
 using Gamesmiths.Forge.Effects;
+using Gamesmiths.Forge.Statescript.Providers;
 
 namespace Gamesmiths.Forge.Statescript.Properties;
