+++ Forge/Statescript/Providers/EventPayloadOutput.cs
@@ -1,6 +1,6 @@
 // Copyright © Gamesmiths Guild.
 
-namespace Gamesmiths.Forge.Statescript;
+namespace Gamesmiths.Forge.Statescript.Providers;
 
 /// <summary>
 /// Declares an output that an <see cref="IEventPayloadProvider"/> exposes to the graph editor when decomposing a
